EvelinLoginPage locators and methods ready
